+++ @navikt/core/react/src/form/file-upload/FileUpload.tsx
@@ -8,27 +8,31 @@ interface FileUploadProps {
 }
 
 interface FileUploadComponent extends React.ExoticComponent<FileUploadProps>{
-  /**
-   * @see 🏷️ {@link DropzoneProps}
-   */
   Dropzone: typeof Dropzone;
-  /**
-   * @see 🏷️ {@link FileListProps}
-   */
   List: typeof List;
-  /**
-   * @see 🏷️ {@link FileItemProps}
-   */
   Item: typeof Item;
 }
 
 /**
- * A component for uploading files. Only used as a namespace. You should
- * use FileUpload.Dropzone instead.
+ * A set of components used to upload files and display files.
+ * FileUpload is only a namespace, only use the subcomponents.
  * @see [📝 Documentation](https://aksel.nav.no/komponenter/core/fileupload)
  * @example
  * ```jsx
- *  <FileUpload.Zone />
+ *  <FileUpload.Dropzone />
+ * ```
+ *
+ * @example
+ * ```jsx
+ *  <FileUpload.List>
+ *    <FileUpload.Item file={myFile} />
+ *    <FileUpload.Item file={mySecondFile} />
+ *  </FileUpload.List>
+ * ```
+ *
+ * @example
+ * ```jsx
+ *  <FileUpload.Item file={myStandaloneFile} />
  * ```
  */
 export const FileUpload = ((props: FileUploadProps) => props.children) as FileUploadComponent;
